[squash-after-review] Move core definitions into CoreBTypes without changes
