+++ src/compiler/scala/tools/nsc/backend/jvm/BTypes.scala
@@ -771,83 +771,4 @@ abstract class BTypes[G <: Global](val __global_dont_use: G) {
       new MethodBType(argTypes, resType)
     }
   }
-
-  val BOXED_UNIT    = ClassBType("java/lang/Void")
-  val BOXED_BOOLEAN = ClassBType("java/lang/Boolean")
-  val BOXED_BYTE    = ClassBType("java/lang/Byte")
-  val BOXED_SHORT   = ClassBType("java/lang/Short")
-  val BOXED_CHAR    = ClassBType("java/lang/Character")
-  val BOXED_INT     = ClassBType("java/lang/Integer")
-  val BOXED_LONG    = ClassBType("java/lang/Long")
-  val BOXED_FLOAT   = ClassBType("java/lang/Float")
-  val BOXED_DOUBLE  = ClassBType("java/lang/Double")
-
-  /*
-   * RT_NOTHING and RT_NULL exist at run-time only. They are the bytecode-level manifestation (in
-   * method signatures only) of what shows up as NothingClass resp. NullClass in Scala ASTs.
-   *
-   * Therefore, when RT_NOTHING or RT_NULL are to be emitted, a mapping is needed: the internal
-   * names of NothingClass and NullClass can't be emitted as-is.
-   */
-  val RT_NOTHING = ClassBType("scala/runtime/Nothing$")
-  val RT_NULL    = ClassBType("scala/runtime/Null$")
-  val CT_NOTHING = ClassBType("scala/Nothing")
-  val CT_NULL    = ClassBType("scala/Null")
-
-  val srBooleanRef = ClassBType("scala/runtime/BooleanRef")
-  val srByteRef    = ClassBType("scala/runtime/ByteRef")
-  val srCharRef    = ClassBType("scala/runtime/CharRef")
-  val srIntRef     = ClassBType("scala/runtime/IntRef")
-  val srLongRef    = ClassBType("scala/runtime/LongRef")
-  val srFloatRef   = ClassBType("scala/runtime/FloatRef")
-  val srDoubleRef  = ClassBType("scala/runtime/DoubleRef")
-
-  /**
-   * Map from type kinds to the Java reference types.
-   * Useful when pushing class literals onto the operand stack (ldc instruction taking a class
-   * literal).
-   * @see Predef.classOf
-   * @see genConstant()
-   *
-   * TODO @lry rename to "boxedClassOfPrimitive" or so, check usages
-   */
-  val classLiteral = immutable.Map[BType, ClassBType](
-    UNIT   -> BOXED_UNIT,
-    BOOL   -> BOXED_BOOLEAN,
-    BYTE   -> BOXED_BYTE,
-    SHORT  -> BOXED_SHORT,
-    CHAR   -> BOXED_CHAR,
-    INT    -> BOXED_INT,
-    LONG   -> BOXED_LONG,
-    FLOAT  -> BOXED_FLOAT,
-    DOUBLE -> BOXED_DOUBLE
-  )
-
-  case class MethodNameAndType(name: String, descriptor: String)
-
-  val asmBoxTo: immutable.Map[BType, MethodNameAndType] = {
-    Map(
-      BOOL   -> MethodNameAndType("boxToBoolean",   "(Z)Ljava/lang/Boolean;"  ) ,
-      BYTE   -> MethodNameAndType("boxToByte",      "(B)Ljava/lang/Byte;"     ) ,
-      CHAR   -> MethodNameAndType("boxToCharacter", "(C)Ljava/lang/Character;") ,
-      SHORT  -> MethodNameAndType("boxToShort",     "(S)Ljava/lang/Short;"    ) ,
-      INT    -> MethodNameAndType("boxToInteger",   "(I)Ljava/lang/Integer;"  ) ,
-      LONG   -> MethodNameAndType("boxToLong",      "(J)Ljava/lang/Long;"     ) ,
-      FLOAT  -> MethodNameAndType("boxToFloat",     "(F)Ljava/lang/Float;"    ) ,
-      DOUBLE -> MethodNameAndType("boxToDouble",    "(D)Ljava/lang/Double;"   )
-    )
-  }
-
-  val asmUnboxTo: immutable.Map[BType, MethodNameAndType] = {
-    Map(
-      BOOL   -> MethodNameAndType("unboxToBoolean", "(Ljava/lang/Object;)Z") ,
-      BYTE   -> MethodNameAndType("unboxToByte",    "(Ljava/lang/Object;)B") ,
-      CHAR   -> MethodNameAndType("unboxToChar",    "(Ljava/lang/Object;)C") ,
-      SHORT  -> MethodNameAndType("unboxToShort",   "(Ljava/lang/Object;)S") ,
-      INT    -> MethodNameAndType("unboxToInt",     "(Ljava/lang/Object;)I") ,
-      LONG   -> MethodNameAndType("unboxToLong",    "(Ljava/lang/Object;)J") ,
-      FLOAT  -> MethodNameAndType("unboxToFloat",   "(Ljava/lang/Object;)F") ,
-      DOUBLE -> MethodNameAndType("unboxToDouble",  "(Ljava/lang/Object;)D")
-    )
-  }
 }
